[DX12] Fix truncation of fence value on playback (uint64 to unsigned)

+++ DirectX/player/replayCustomizationLayer.cpp
@@ -1948,7 +1948,7 @@ void ReplayCustomizationLayer::FillCpuDescriptorHandleArgument(
 
 void ReplayCustomizationLayer::WaitForFence(unsigned commandKey,
                                             ID3D12Fence* fence,
-                                            unsigned fenceValue) {
+                                            UINT64 fenceValue) {
   UINT64 value = fence->GetCompletedValue();
   if (value >= fenceValue) {
     return;
